+++ example/src/index.ts
@@ -14,10 +14,26 @@ const app = new Spooked({
         password: process.env.DB_PASS || '',
         name: process.env.DB_NAME || 'Spooked',
     },
+    server: {
+        host: process.env.SERVER_HOST || 'localhost',
+        port: Number(process.env.SERVER_PORT) || 5000,
+    },
 });
 
+async function runApp(): Promise<void> {
+    const database = app.getDatabase();
+
+    // Connect to the database.
+    await database.connect();
+
+    const server = app.getServer();
+
+    // Initialize the server.
+    await server.listen();
+}
+
 try {
-    app.getDatabaseManager().connect();
+    runApp();
 } catch (e) {
     console.log('Spooked Error', e);
 }